product added

+++ src/main/java/com/stan/stancommerce/entities/Category.java
@@ -17,7 +17,8 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Byte id;
+    //private Byte id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -29,7 +30,10 @@ public class Category {
         this.name = name;
     }
 
-    public Category(byte id) {
+//   // public Category(byte id) {
+//        this.id = id;
+//    }
+    public Category(long id) {
         this.id = id;
     }
 }
